pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pcgasmeter"
description = 'A small package to read the red gas volume dial of a Parkinson-Cowan gas meter from camera images.'
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "Ulrich Dobramysl", email = "" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "Pillow>=9.2.0",
  "scikit-image>=0.19.3",
  "opencv-python>=4.6.0",
  "numpy>=1.23.2",
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/ulido/pcgasmeter#readme"
Issues = "https://github.com/ulido/pcgasmeter/issues"
Source = "https://github.com/ulido/pcgasmeter"

[tool.hatch.version]
path = "pcgasmeter/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=pcgasmeter --cov=tests"
no-cov = "cov --no-cov"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "pcgasmeter/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
